\name{coef-methods}
\docType{methods}
\alias{coef-methods}
\alias{coef,ANY-method}
\alias{coef,glmgen-method}
\title{Coefficients -- S4 Methods}
\description{
  The \code{"glmgen"} package provides methods for
  calculating the coefficients of a generalized lasso
  fit via the generic function \code{coef}.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "ANY")}}{

}

\item{\code{signature(object = "glmgen")}}{

}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
