% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.glmgen}
\alias{coef.glmgen}
\title{Get coefficients from a glmgen object}
\usage{
\method{coef}{glmgen}(object, lambda = NULL, ...)
}
\arguments{
\item{object}{output of glmgen function}

\item{lambda}{optional vector of lambda values to calculate coefficients
at. If missing, will use break points in the fit.}

\item{...}{optional, currently unused, arguments}
}
\description{
Get coefficients from a glmgen object
}

