% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.trendfilter}
\alias{predict.trendfilter}
\title{Get predictions from a trendfilter object}
\usage{
\method{predict}{trendfilter}(obj, lambdas = NULL, x.new = NULL, zero_tol = 1e-06)
}
\arguments{
\item{obj}{Object of class \code{trendfilter}.}

\item{lambdas}{(Optional) Vector of lambda values to calculate coefficients
at. If missing, will use break points in the fit.}

\item{x.new}{Vector of new \code{x} points. Set to \code{NULL} (the default) to use the
original locations.}

\item{zero_tol}{Numerical tolerance parameter, for determining whether a}

\item{type}{Scale of the predictions.}
}
\description{
Get predictions from a trendfilter object
}
