% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{print.glmgen}
\alias{print.glmgen}
\title{Print the output of a glmgen object}
\usage{
\method{print}{glmgen}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{optional, currently unused, arguments}
}
\description{
Print the output of a glmgen object
}

