% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.glmgen}
\alias{summary.glmgen}
\title{Summarize a generic glmgen object}
\usage{
\method{summary}{glmgen}(object, ...)
}
\arguments{
\item{object}{object to summarize}

\item{...}{optional, currently unused, arguments}
}
\description{
Summarize a generic glmgen object
}
