% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendfilter.R
\name{tfMultiply}
\alias{tfMultiply}
\title{Multiply a vector by trendfilter matricies}
\usage{
tfMultiply(x, y, k = 2L)
}
\arguments{
\item{x}{Vector of observed inputs}

\item{y}{Numeric vector to which the multiplication should be supplied.}

\item{k}{Order of the trendfiltering matrix.}
}
\value{
A numeric vector with the result of the multiplication.
}
\description{
Multiply a vector by trendfilter matricies
}
\seealso{
\code{\link[=trendfilter]{trendfilter()}}
}
