% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.trendfilter}
\alias{predict.trendfilter}
\title{Get predictions from a trendfilter object}
\usage{
\method{predict}{trendfilter}(object, type = c("link", "response"),
  lambda = NULL, x.new = NULL, zero_tol = 1e-06, ...)
}
\arguments{
\item{object}{output of trendfilter function}

\item{type}{scale of the predictions}

\item{lambda}{optional vector of lambda values to calculate coefficients
at. If missing, will use break points in the fit.}

\item{x.new}{vector of new x points. Set to NULL (the default) to use the
original locations.}

\item{zero_tol}{numerical tolerance parameter, for determining whether a}

\item{...}{optional, currently unused, arguments}
}
\description{
Get predictions from a trendfilter object
}
