% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.trendfilter}
\alias{predict.trendfilter}
\title{Get predictions from a trendfilter object}
\usage{
\method{predict}{trendfilter}(object, lambda = NULL, ...)
}
\arguments{
\item{object}{output of summary.iolm}

\item{lambda}{optional vector of lambda values to calculate coefficients
at. If missing, will use break points in the fit.}

\item{...}{optional, currently unused, arguments}
}
\description{
Get predictions from a trendfilter object
}

