% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.glmgen}
\alias{print.summary.glmgen}
\title{Print the output of a glmgen summary object}
\usage{
\method{print}{summary.glmgen}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{optional, currently unused, arguments}
}
\description{
Print the output of a glmgen summary object
}
