% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.trendfilter}
\alias{summary.trendfilter}
\title{Summarize a trendfilter glmgen object}
\usage{
\method{summary}{trendfilter}(object, ...)
}
\arguments{
\item{object}{object to summarize}

\item{...}{optional, currently unused, arguments}
}
\description{
Summarize a trendfilter glmgen object
}

