% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendfilter.R
\name{tfMultiply}
\alias{tfMultiply}
\title{Multiply a vector by Trendfilter Matricies}
\usage{
tfMultiply(b, k = 1L, x = NULL, matrix = c("d"))
}
\arguments{
\item{b}{a numeric vector to which the multiplication should
be supplied}

\item{k}{a positive integer; the order of the trendfiltering matrix}

\item{x}{the positions associated with the matrix. If set to NULL,
these will be automatically set to \code{1:length(b)}}

\item{matrix}{option for which matrix should be used for multiplication.
See Details for more information.}
}
\value{
a numeric vector with the result of the multiplication
}
\description{
Fast algorithms exist for multiplying a vector by
the trendfiltering penalty matrix D and other
related matricies, and are used internally in the
function \code{\link{trendfilter}}. Here we expose
functions for advanced users to
}
\examples{
 set.seed(0)
 y <- rnorm(100)
 all.equal(diff(y), tfMultiply(y))

}
\references{
Tibshirani, R. J. (2014), "Adaptive piecewise polynomial estimation
    via trend filtering", Annals of Statistics 42 (1): 285--323.

  Ramdas, A. and Tibshirani R. J. (2014), "Fast and flexible ADMM algorithms
    for trend filtering", arXiv: 1406.2082.
}
\seealso{
\code{\link{trendfilter}}
}
\author{
Taylor Arnold, Veeranjaneyulu Sadhanala, Ryan Tibshirani
}
